some tiny misprints, warnings and style issues fixed + PathCoverWeightCounter constructor simplified + major refactoring of meta exspander launch: metagenomic-specific classes and factory methods removed as obsolete ; everything done with proper configuration

+++ assembler/src/common/modules/path_extend/pe_io.hpp
@@ -88,7 +88,7 @@ protected:
 
                     overlapLen = overlapLen + (int) k_ - path.GapAt(i);
                     if(overlapLen < 0) {
-                        for (size_t j = 0; j < abs(overlapLen); ++j) {
+                        for (int j = 0; j < abs(overlapLen); ++j) {
                             ss << "N";
                         }
                         overlapLen = 0;
